Moving back blog post to pages
