+++ package.json
@@ -22,7 +22,7 @@
     "caniuse-lite": "^1.0.30001489",
     "date-fns": "^2.28.0",
     "gray-matter": "^4.0.3",
-    "next": "13.4.3",
+    "next": "13.4.4",
     "next-mdx-remote": "^4.0.2",
     "openai": "^3.2.1",
     "react": "18.2.0",
